+++ packages/editor/src/components/collaborators-presence/avatar-group/component.tsx
@@ -7,19 +7,20 @@ import clsx from 'clsx';
  * WordPress dependencies
  */
 import { Children } from '@wordpress/element';
+import { sprintf, _n } from '@wordpress/i18n';
 
 /**
  * Internal dependencies
  */
 import type { AvatarGroupProps } from './types';
-import type { WordPressComponentProps } from '../context';
 
 function AvatarGroup( {
 	className,
 	max = 3,
 	children,
 	...props
-}: WordPressComponentProps< AvatarGroupProps, 'div', false > ) {
+}: AvatarGroupProps &
+	Omit< React.HTMLAttributes< HTMLDivElement >, keyof AvatarGroupProps > ) {
 	const childArray = Children.toArray( children );
 	const visible = childArray.slice( 0, max );
 	const overflowCount = childArray.length - max;
@@ -27,14 +28,22 @@ function AvatarGroup( {
 	return (
 		<div
 			role="group"
-			className={ clsx( 'components-avatar-group', className ) }
+			className={ clsx( 'editor-avatar-group', className ) }
 			{ ...props }
 		>
 			{ visible }
 			{ overflowCount > 0 && (
 				<span
-					className="components-avatar-group__overflow"
-					aria-label={ `${ overflowCount } more` }
+					className="editor-avatar-group__overflow"
+					aria-label={ sprintf(
+						/* translators: %d: number of additional collaborators not shown */
+						_n(
+							'%d more collaborator',
+							'%d more collaborators',
+							overflowCount
+						),
+						overflowCount
+					) }
 				>
 					{ `+${ overflowCount }` }
 				</span>